After Effects: パラメータ追加共通化とカテゴリ統一
CParamsSetupクラスを新規追加し、各プラグインのParamsSetup処理を共通化・簡略化しました。AlphaFixやPluginSkeleton等でパラメータ追加コードをCParamsSetup経由に書き換え、ボタンやトピック等のサンプルも追加。ピクセル処理もC++20テンプレート化で共通化し、型安全性を向上。全プラグインのカテゴリを "NF's Plugins-{Legacy}" に統一し、PiPLリソースも一括修正。ビルド設定のC++20化や警告抑制、細かなUI・ID修正も含みます。

+++ InnerFlare/Tiny_AlphaHyperbolic.cpp
@@ -51,7 +51,7 @@ static PF_Err HyperbolicAlpha(
     }
 
     // ハイパーボリック処理
-    float k = infoP->hyperbolic;
+    float k = (float)infoP->hyperbolic;
     float v = ((1.0f + k) * t) / (1.0f + k * t);
 
     // 範囲チェック
